requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
aiofiles==23.2.1
itsdangerous
argon2-cffi
